Changed tests of auth screens
