+++ src/tests/containers/Login/page.test.js
@@ -1,7 +1,24 @@
+import React from 'react';
+
+import { render } from '@testing-library/react';
+
 import Page from 'containers/Login/page';
-import AuthPage from 'tests/shared_examples/authPage';
+import backgroundImage from 'images/auth_background.png';
+
 
 describe('When visited the Login Page', () => {
   const logoLabel = 'Sign in to Nimble';
-  AuthPage(logoLabel, Page);
+  it('shows a background image', () => {
+    const { getByLabelText } = render(<Page />);
+    const imageDiv = getByLabelText('background-layout-image');
+
+    expect(imageDiv).toHaveStyle(`background-image: url(${backgroundImage})`);
+  });
+
+  it('shows a logo label', () => {
+    const { getByText } = render(<Page />);
+    const labelText = getByText(logoLabel);
+
+    expect(labelText).toBeInTheDocument();
+  });
 });
